+++ src/robot_upstart/job.py
@@ -39,7 +39,7 @@ import providers
 class Job(object):
     """ Represents a ROS configuration to launch on machine startup. """
 
-    def __init__(self, name="ros", interface=None, user=None, groups=None, workspace_setup=None,
+    def __init__(self, name="ros", interface=None, user=None, groups=None, envfile=None, workspace_setup=None,
                  rosdistro=None, master_uri=None, log_path=None):
         """Construct a new Job definition.
 
@@ -54,6 +54,10 @@ class Job(object):
         :param user: Unprivileged user to launch the job as. Defaults to the user
             creating the job.
         :type user: str
+        :param groups: Specify supplementary groups (comma separated) for command execution (will use runuser with -G option)
+        :type groups: str
+        :param envfile: environment file (valid with systemd provider)
+        :type envfile: str
         :param workspace_setup: Location of the workspace setup file to source for
             the job's ROS context. Defaults to the current workspace.
         :type workspace_setup: str
@@ -77,6 +81,8 @@ class Job(object):
         self.user = user or getpass.getuser()
 
         self.groups = groups
+
+        self.envfile = envfile
 
         # Fall back on current workspace setup file if not explicitly specified.
         self.workspace_setup = workspace_setup or \